renaming, refactor TurnExpert
